requirements.txt
asgiref==3.2.10
certifi==2020.6.20
cffi==1.14.2
chardet==3.0.4
cryptography==3.1
decorator==4.4.2
Django==3.1.1
django-crispy-forms==1.9.2
django-rest-knox==4.1.0
djangorestframework==3.11.1
idna==2.10
pycparser==2.20
python-telegram-bot==12.8
pytz==2020.1
requests==2.24.0
six==1.15.0
sqlparse==0.3.1
tornado==6.0.4
urllib3==1.25.10
